+++ xlsxreport/scripts/cli.py
@@ -3,6 +3,7 @@ import click
 
 from xlsxreport.scripts.appdir import appdir_command
 from xlsxreport.scripts.compile_excel import compile_excel_command
+from xlsxreport.scripts.validate import validate_command
 
 
 @click.group()
@@ -13,6 +14,7 @@ def cli():
 
 cli.add_command(appdir_command, name="appdir")
 cli.add_command(compile_excel_command, name="compile")
+cli.add_command(validate_command, name="validate")
 
 if __name__ == "__main__":
     cli()